added error logging support
The frame work is in place to log errors to error.log. Error logging is capable of synchronous error logging from server.py, telemetry.py and aprs.py.  Currently no error messages are in place, just the functionality to use them.  Next step is identifying what errors there could be and what we want to log when they arise